Moved WeatherTestUtils and linked where code in build.gradle tests comes from stackoverflow
