+++ build.gradle
@@ -42,6 +42,8 @@ tasks.named('test') {
 	}
 	afterSuite { desc, result -> // prints the results of the test
 		if (!desc.parent) { // will match the outermost suite
+
+			// Printing the results of the test - from https://stackoverflow.com/questions/3963708/gradle-how-to-display-test-results-in-the-console-in-real-time
 			def output = "Results: ${result.resultType} (${result.testCount} tests, ${result.successfulTestCount} passed, ${result.failedTestCount} failed, ${result.skippedTestCount} skipped)"
 			def startItem = '|  ', endItem = '  |'
 			def repeatLength = startItem.length() + output.length() + endItem.length()
