+++ src/test/java/com/example/weatherapi/api/SmhiApiTests.java
@@ -9,7 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 
-import static com.example.weatherapi.util.WeatherTestUtils.*;
+import static com.example.weatherapi.utilitytests.WeatherTestUtils.*;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @ActiveProfiles("test")
