+++ src/test/java/com/example/weatherapi/controllers/CityEntityControllerTests.java
@@ -15,7 +15,6 @@ import org.springframework.test.context.ActiveProfiles;
 import java.time.OffsetDateTime;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.offset;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @ActiveProfiles("test")
